combine / put things back
